pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Andrei Emelianenko <spliitvein@gmail.com>"]
readme = "README.md"
packages = [{include = "task_manager"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.4"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
dj-database-url = "^2.1.0"
whitenoise = {extras = ["brotil"], version = "^6.6.0"}
gunicorn = "^22.0.0"
uvicorn = "^0.29.0"
django-bootstrap5 = "^24.2"
django-filter = "^24.2"
rollbar = "^1.0.0"
pytest = "^8.2.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
coverage = "^7.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
